Fix SSH reconnect guard failures
