+++ cmuxTests/SSHStartupSignalLifecycleTests.swift
@@ -552,84 +552,7 @@ extension CLINotifyProcessIntegrationRegressionTests {
         XCTAssertFalse(result.timedOut, result.stderr)
         XCTAssertEqual(result.status, 1, result.stderr)
         XCTAssertTrue(result.stderr.contains("[cmux] ssh exited with status 1."), result.stderr)
-        XCTAssertTrue(
-            result.stderr.contains("[cmux] press r then Enter to reconnect, or Enter to close this pane."),
-            result.stderr
-        )
-        // EOF on the hold prompt (no stdin) dismisses: session-end cleanup runs exactly once.
-        let dismissSessionEndCalls = ((try? String(contentsOf: logFile, encoding: .utf8)) ?? "")
-            .split(separator: "\n")
-            .filter { $0.contains("ssh-session-end") }
-        XCTAssertEqual(dismissSessionEndCalls.count, 1, result.stderr)
-    }
-
-    func testSSHStartupManualReconnectReentersConnectLoop() throws {
-        // When the automatic reconnect loop gives up on a non-retryable exit, the hold
-        // prompt lets the user press `r` (then Enter) to re-run the SSH connect loop from
-        // scratch instead of closing the pane. Here the fake ssh fails on the first pass
-        // and succeeds on the second, so a single `r` must produce a clean (status 0)
-        // session with the transport session-end reported exactly once (on final exit).
-        let fileManager = FileManager.default
-        let root = fileManager.temporaryDirectory
-            .appendingPathComponent("cmux-ssh-manual-retry-\(UUID().uuidString)", isDirectory: true)
-        let fakeCLI = root.appendingPathComponent("cmux")
-        let fakeSSH = root.appendingPathComponent("ssh")
-        let logFile = root.appendingPathComponent("ssh-session-end.log")
-        let attemptFile = root.appendingPathComponent("ssh-attempts.txt")
-
-        try fileManager.createDirectory(at: root, withIntermediateDirectories: true)
-        defer { try? fileManager.removeItem(at: root) }
-
-        try writeShellFile(at: fakeCLI, lines: [
-            "#!/bin/sh",
-            "printf '%s\\n' \"$*\" >> \"${CMUX_TEST_SESSION_END_LOG}\"",
-        ])
-        // Fail on the first invocation (status 1, non-retryable so the inner loop stops
-        // immediately and reaches the hold prompt), then succeed on the second so the
-        // manual retry produces a clean session.
-        try writeShellFile(at: fakeSSH, lines: [
-            "#!/bin/sh",
-            "count=$(cat \"${CMUX_TEST_ATTEMPT_FILE}\" 2>/dev/null || printf 0)",
-            "count=$((count + 1))",
-            "printf '%s' \"$count\" > \"${CMUX_TEST_ATTEMPT_FILE}\"",
-            "if [ \"$count\" -ge 2 ]; then exit 0; fi",
-            "exit 1",
-        ])
-        try fileManager.setAttributes([.posixPermissions: 0o700], ofItemAtPath: fakeCLI.path)
-        try fileManager.setAttributes([.posixPermissions: 0o700], ofItemAtPath: fakeSSH.path)
-
-        let startupCommand = try generatedVMSSHInitialStartupCommand()
-        var environment = ProcessInfo.processInfo.environment
-        environment["PATH"] = "\(root.path):\(environment["PATH"] ?? "/usr/bin:/bin")"
-        environment["CMUX_BUNDLED_CLI_PATH"] = fakeCLI.path
-        environment["CMUX_SOCKET_PATH"] = "/tmp/cmux-debug-test.sock"
-        environment["CMUX_WORKSPACE_ID"] = "11111111-1111-1111-1111-111111111111"
-        environment["CMUX_SURFACE_ID"] = "22222222-2222-2222-2222-222222222222"
-        environment["CMUX_TEST_SESSION_END_LOG"] = logFile.path
-        environment["CMUX_TEST_ATTEMPT_FILE"] = attemptFile.path
-        environment["CMUX_SSH_RECONNECT_DELAY_SECONDS"] = "0"
-
-        let result = runProcess(
-            executablePath: "/bin/sh",
-            arguments: ["-c", startupCommand],
-            environment: environment,
-            standardInput: "r\n",
-            timeout: 5
-        )
-
-        XCTAssertFalse(result.timedOut, result.stderr)
-        XCTAssertEqual(result.status, 0, result.stderr)
-        XCTAssertEqual(
-            (try? String(contentsOf: attemptFile, encoding: .utf8))?.trimmingCharacters(in: .whitespacesAndNewlines),
-            "2",
-            "manual `r` retry must re-run the ssh connect loop a second time"
-        )
-        // Manual retry must not report session-end between the failed and successful
-        // passes; cleanup runs exactly once, on the final clean exit.
-        let sessionEndCalls = ((try? String(contentsOf: logFile, encoding: .utf8)) ?? "")
-            .split(separator: "\n")
-            .filter { $0.contains("ssh-session-end") }
-        XCTAssertEqual(sessionEndCalls.count, 1, result.stderr)
+        XCTAssertTrue(result.stderr.contains("[cmux] press r then Enter to reconnect, or Enter to close this pane."), result.stderr)
     }
 
     func testSSHStartupForwardsStdinToBackgroundedSSH() throws {
@@ -795,7 +718,7 @@ extension CLINotifyProcessIntegrationRegressionTests {
         return try XCTUnwrap(configureParams["terminal_startup_command"] as? String)
     }
 
-    private func generatedVMSSHInitialStartupCommand() throws -> String {
+    func generatedVMSSHInitialStartupCommand() throws -> String {
         let cliPath = try bundledCLIPath()
         let socketPath = makeSocketPath("vm-ssh-startup")
         let listenerFD = try bindUnixSocket(at: socketPath)
@@ -897,7 +820,7 @@ extension CLINotifyProcessIntegrationRegressionTests {
         return try XCTUnwrap(createParams["initial_command"] as? String)
     }
 
-    private func writeShellFile(at url: URL, lines: [String]) throws {
+    func writeShellFile(at url: URL, lines: [String]) throws {
         try lines.joined(separator: "\n")
             .appending("\n")
             .write(to: url, atomically: true, encoding: .utf8)
